feat: allow SDK to trigger avatar emotes
