+++ kernel/packages/shared/apis/index.ts
@@ -7,4 +7,4 @@ import './SocialController'
 import './DevTools'
 import './CommunicationsController'
 import './UserActionModule'
-import './RestrictedActionModule'
+import './RestrictedActions'
