+++ kernel/public/ecs-scenes/0.101.move-inside-scene/scene.json
