+++ kernel/public/ecs-scenes/0.101.move-inside-scene/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+      "outFile": "./game.js"
+    },
+    "include": ["*.ts"],
+    "extends": "../../../packages/decentraland-ecs/types/tsconfig.json"
+}
