+++ kernel/public/ecs-scenes/0.104.trigger-emotes/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+      "outFile": "./game.js"
+    },
+    "include": ["*.ts"],
+    "extends": "../../../packages/decentraland-ecs/types/tsconfig.json"
+}
