+++ kernel/test/index.ts
@@ -20,7 +20,7 @@ import './unit/jsonFetch.test'
 import './unit/profiles.saga.test'
 import './unit/BrowserInterface.test'
 import './unit/positionThings.test'
-import './unit/RestrictedActionModule.test'
+import './unit/RestrictedActions.test'
 import './unityIntegration/ecs/math/quaternion.test'
 import './unityIntegration/ecs/math/vector3.test'
